refactor: 게시물 get 파일 분리
